+++ .github/workflows/test.yaml
@@ -92,9 +92,9 @@ jobs:
       - name: Combine coverage
         run: poetry run coverage combine coverage
       - name: Coverage report
-        run: poetry run coverage report --skip-empty --show-missing --include=oas_tools/*
+        run: poetry run coverage report --skip-empty --show-missing --include=openapi_spec_tools/*
       - name: Coverage HTML
-        run: poetry run coverage html --show-contexts --title "Coverage for ${{ github.sha }}" --include=oas_tools/*
+        run: poetry run coverage html --show-contexts --title "Coverage for ${{ github.sha }}" --include=openapi_spec_tools/*
       - name: Store coverage HTML
         uses: actions/upload-artifact@v4
         with:
