+++ openapi_spec_tools/cli_gen/_console.py
