+++ openapi_spec_tools/cli_gen/_exceptions.py
@@ -1,7 +1,7 @@
 import typer
 from requests import HTTPError
 
-from oas_tools.cli_gen._console import console_factory
+from openapi_spec_tools.cli_gen._console import console_factory
 
 
 class MissingRequiredError(Exception):
