+++ openapi_spec_tools/cli_gen/constants.py
