+++ openapi_spec_tools/cli_gen/utils.py
