+++ openapi_spec_tools/utils.py
